--- server/controllers/index.js
@@ -1,12 +1,0 @@
-export { default as signupController } from "./auth/signupController";
-export { default as loginController } from "./auth/loginController";
-export { default as emailVerificationController } from "./auth/emailVerificationController";
-export { default as changePasswordController } from "./auth/changePasswordController";
-export { default as refreshTokenController } from "./auth/refreshTokenController";
-
-export { default as aboutController } from "./profile/aboutController";
-export { default as friendlistController } from "./profile/friendlistController";
-
-export { default as friendRequestController } from "./friendRequest/friendRequestController";
-
-export { default as postController } from "./post/postController";
