+++ server/env_sample
@@ -1,8 +1,12 @@
+SERVER_URL=SERVER_URL
+CLIENT_URL=CLIENT_URL
 DEV_PORT=PORT_NUMBER
 DB_URL=MongoDB_URL
 
 JWT_ACCESS_SECRET=1234
 JWT_REFRESH_SECRET=123456
+ACCESS_TOKEN_EXIRY=EXPIRY
+REFRESH_TOKEN_EXIRY=EXPIRY
 
 # REFER https://www.youtube.com/watch?v=lBRnLXwjLw0
