--- server/models/index.js
@@ -1,4 +1,0 @@
-export { default as UserModel } from "./userModel";
-export { default as PostModel } from "./postModel";
-export { default as RefreshTokenModel } from "./refreshTokenModel";
-export { default as EmailTokenModel } from "./emailTokenModel";
