+++ server/services/index.js
@@ -1,2 +1,1 @@
 export { default as customErrorHandler } from "./customErrorHandler";
-export { default as jwtService } from "./jwtService";
